Initial POC for rabbitmq support
- Added rabbitmq server and reciever app to dev environment
- Starting dev environment with `--rabbitmq` will sent env var `CONJUR_ENABLE_RABBITMQ=true` in conjur
- The `rabbitmq-reciever` application is a rather complete example of how we might expect end-users to use this live-notification feature
- RabbitMQ connection singleton added conjur, but no publishing has been implemented yet

Also want to add a "rabbitmq" policy that this class will read upon server boot. It would parse the policy for annotations of groups/layers of which to create queues for. Perhaps each should have their own credentials to ensure that only the target host(s) can connect to the queue and recieve a stream of secrets being updated.

TODO:
- In conjur, when secret value changes/is rotated, send message to hosts that can read+execute that secret (targeting a queue sharing the same name as a group/layer)
- When a new group/layer is added to the policy, I think conjur will have to restart... so we should probably have some way to indicate queues opened by conjur